chore: delete utils
